+++ src/test/java/com/rabbitmq/stream/impl/BatchEntryTest.java
@@ -65,7 +65,7 @@ public class BatchEntryTest {
                     consumeLatch.countDown();
                 }));
 
-        Client.Response response = consumer.subscribe(1, stream, OffsetSpecification.first(), 10);
+        Client.Response response = consumer.subscribe((byte) 1, stream, OffsetSpecification.first(), 10);
         assertThat(response.isOk()).isTrue();
 
         assertThat(consumeLatch.await(10, TimeUnit.SECONDS)).isTrue();
